+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Changes
 =======
 
-TypeTree 2.9-git
-----------------
+TypeTree 2.9
+------------
 
 -   Remove `apply_to_tuple` utility.
 -   Add support for `ProxyNode` on nodes with dynamic degree.